+++ src/anvil/api/features.py
@@ -2,15 +2,18 @@ import json
 import os
 from enum import StrEnum
 
+from griffe import Logger
+from PIL import Image, ImageDraw, ImageFont
+
 from anvil import ANVIL, CONFIG
-from anvil.api.enums import (CameraPresets, FogCameraLocation, LootPoolType,
-                             PotionId, RawTextConstructor, RenderDistanceType,
+from anvil.api.enums import (AimAssistTargetMode, CameraPresets,
+                             FogCameraLocation, LootPoolType, PotionId,
+                             RawTextConstructor, RenderDistanceType,
                              SmeltingTags)
 from anvil.api.types import Identifier
 from anvil.lib.lib import CopyFiles, File, FileExists, clamp
 from anvil.lib.reports import ReportType
 from anvil.lib.schemas import AddonObject, JsonSchemes, MinecraftDescription
-from PIL import Image, ImageDraw, ImageFont
 
 
 # Dialogue ------------------------------------------------
@@ -851,6 +854,99 @@ class Particle(AddonObject):
 
 
 # Camera Presets ------------------------------------------
+
+
+class AimAssistPreset(AddonObject):
+    _extension = ".json"
+    _path = os.path.join(CONFIG.BP_PATH, "Camera", "Presets")
+
+    def __init__(self, identifier: str):
+        super().__init__(identifier)
+        self.content(JsonSchemes.aim_assist_preset(identifier))
+
+    @property
+    def identifier(self):
+        return self._content["minecraft:aim_assist_preset"]["identifier"]
+
+    def item_settings(self, settings: dict):
+        self._preset["minecraft:aim_assist_preset"]["item_settings"] = settings
+        return self
+
+    def default_item_settings(self, setting: str):
+        self._preset["minecraft:aim_assist_preset"]["default_item_settings"] = setting
+        return self
+
+    def hand_settings(self, setting: str):
+        self._preset["minecraft:aim_assist_preset"]["hand_settings"] = setting
+        return self
+
+    def exclusion_list(self, exclusions: dict):
+        self._preset["minecraft:aim_assist_preset"]["exclusion_list"] = exclusions
+        return self
+
+    def liquid_targeting_list(self, targets: dict):
+        self._preset["minecraft:aim_assist_preset"]["liquid_targeting_list"] = targets
+        return self
+
+    def _export(self):
+        return self._preset
+
+    def queue(self, directory: str = None):
+        self.content(self._preset)
+        return super().queue(directory)
+
+
+class AimAssistCategory:
+    def __init__(self, name: str):
+        self._category = {"name": name, "priorities": {}}
+
+    def entity_default(self, value: int):
+        self._category["entity_default"] = value
+        return self
+
+    def block_default(self, value: int):
+        self._category["block_default"] = value
+        return self
+
+    def block_priority(self, block: str, priority: int):
+        if "blocks" not in self._category["priorities"]:
+            self._category["priorities"]["blocks"] = {}
+        self._category["priorities"]["blocks"][block] = priority
+        return self
+
+    def entity_priority(self, entity: str, priority: int):
+        if "entities" not in self._category["priorities"]:
+            self._category["priorities"]["entities"] = {}
+        self._category["priorities"]["entities"][entity] = priority
+        return self
+
+    def export(self):
+        return self._category
+
+
+class AimAssistCategories(AddonObject):
+    _extension = ".json"
+    _path = os.path.join(CONFIG.BP_PATH, "Camera", "Presets")
+
+    def __init__(self):
+        super().__init__("categories")
+        self.content(JsonSchemes.aim_assist_categories())
+        self._categories: list[AimAssistCategory] = []
+
+    def add_category(self, category_name: str):
+        category = AimAssistCategory(category_name)
+        self._categories.append(category)
+        return category
+
+    def _export(self):
+        for category in self._categories:
+            self._content["minecraft:aim_assist_categories"]["categories"].append(category.export())
+        return self._content
+
+    def queue(self):
+        return super().queue()
+
+
 class CameraPreset(AddonObject):
     """A class representing a CameraPreset."""
 
@@ -868,6 +964,7 @@ class CameraPreset(AddonObject):
         self._name = name
         self._inherit = inherit_from.value if isinstance(inherit_from, CameraPresets) else str(inherit_from)
         self._camera_preset = JsonSchemes.camera_preset(CONFIG.NAMESPACE, name, self._inherit)
+        self._replace_reticle = False
 
     def position(self, x: float = 0, y: float = 0, z: float = 0):
         """Sets the position of the camera preset.
@@ -933,19 +1030,79 @@ class CameraPreset(AddonObject):
     def view_offset(self, x_offset: float, y_offset: float):
         self._camera_preset["minecraft:camera_preset"]["view_offset"] = [x_offset, y_offset]
         return self
-    
+
     def entity_offset(self, x_offset: float, y_offset: float, z_offset: float):
         self._camera_preset["minecraft:camera_preset"]["entity_offset"] = [x_offset, y_offset, z_offset]
         return self
-    
+
     def radius(self, radius: float):
         self._camera_preset["minecraft:camera_preset"]["radius"] = clamp(radius, 0.1, 100)
+        return self
+
+    def aim_assist(
+        self,
+        preset: AimAssistPreset,
+        target_mode: AimAssistTargetMode = AimAssistTargetMode.Distance,
+        angle: list[float] = [30, 30],
+        distance: float = 8,
+        replace_reticle: bool = False,
+    ):
+        self._camera_preset["minecraft:camera_preset"]["aim_assist"] = {
+            "preset": preset.identifier,
+            "target_mode": target_mode.value,
+            "angle": [min(angle), max(angle)],
+            "distance": clamp(distance, 1, 16),
+        }
+
+        if replace_reticle:
+            self._replace_reticle = True
+        return self
+
+    def focus_target(
+        self,
+        rotation_speed: float = 0.0,
+        snap_to_target: bool = False,
+        horizontal_rotation_limit: list[float] = [0, 360],
+        vertical_rotation_limit: list[float] = [0, 180],
+        continue_targeting: bool = False,
+        tracking_radius: float = 50.0,
+    ):
+        self._camera_preset["minecraft:camera_preset"]["rotation_speed"] = max(0.0, rotation_speed)
+        self._camera_preset["minecraft:camera_preset"]["snap_to_target"] = snap_to_target
+
+        self._camera_preset["minecraft:camera_preset"]["horizontal_rotation_limit"] = [
+            max(horizontal_rotation_limit[0], 0),
+            min(horizontal_rotation_limit[1], 360),
+        ]
+        self._camera_preset["minecraft:camera_preset"]["vertical_rotation_limit"] = [
+            max(vertical_rotation_limit[0], 0),
+            min(vertical_rotation_limit[1], 180),
+        ]
+        self._camera_preset["minecraft:camera_preset"]["continue_targeting"] = continue_targeting
+        self._camera_preset["minecraft:camera_preset"]["tracking_radius"] = tracking_radius
         return self
 
     def queue(self):
         """Queues the camera preset to be exported."""
         self.content(self._camera_preset)
         return super().queue()
+
+    def _export(self):
+        if self._replace_reticle:
+            if FileExists(os.path.join("assets", "textures", "ui", "aimassist_block_highlight.png")):
+                CopyFiles(
+                    os.path.join("assets", "textures", "ui"),
+                    os.path.join(CONFIG.RP_PATH, "textures", "ui"),
+                    "aimassist_block_highlight.png",
+                )
+            if FileExists(os.path.join("assets", "textures", "ui", "aimassist_entity_highlight.png")):
+                CopyFiles(
+                    os.path.join("assets", "textures", "ui"),
+                    os.path.join(CONFIG.RP_PATH, "textures", "ui"),
+                    "aimassist_entity_highlight.png",
+                )
+
+        return super()._export()
 
     def __str__(self) -> str:
         return f"{CONFIG.NAMESPACE}:{self._name}"